backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
anthropic==0.18.0
redis==5.0.0
pdf2image==1.17.0
Pillow==10.2.0
pydantic==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.9
httpx==0.26.0
